+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## v0.13.0 (August 18, 2016)
 * *[v0.13.0-rc4]*
   * [FIX]: Fixed line-height issue on account summary
+  * [FEATURE]: Refresh the campaign/analytics data every 5 seconds in
+    the campaign detail view
+  * [FIX]: Make sure the user sees how many views their campaign will
+    get after creating it
 
 * *[v0.13.0-rc3]*
   * Update interface styles